changed Int32 to Int64 when reading a mobile number to avoid overflow.

+++ CashbackCalculator.aspx.cs
@@ -26,7 +26,7 @@ namespace Telecom_Company_Team_9
             string data = "EXEC Payment_wallet_cashback @mobile_num = @mobileNo, @payment_id = @paymentID, @benefit_id = @benefitID";
 
             // Get the mobile number from the input field
-            Int32 mobileNumber = int.Parse(InputMobileNumber.Text);
+            Int64 mobileNumber = int.Parse(InputMobileNumber.Text);
             Int32 paymentID = int.Parse(InputPaymentID.Text);
             Int32 benefitID = int.Parse(InputBenefitID.Text);
